+++ packages/cursorless-engine/src/processTargets/modifiers/scopeHandlers/TreeSitterScopeHandler/index.ts
@@ -1,0 +1,1 @@
+export * from "./TreeSitterScopeHandler";
